moving proactive event handler to be with changeNotificationSettingsIntentHandler
