+++ lambda/custom/handlers/helperIntents.js
@@ -1,30 +1,6 @@
 const { ri } = require('@jargon/alexa-skill-sdk');
 const { customLog } = require('../helperFunctions');
 
-const ProactiveEventHandler = {
-	canHandle(handlerInput) {
-		return handlerInput.requestEnvelope.request.type === 'AlexaSkillEvent.ProactiveSubscriptionChanged';
-	},
-	async handle(handlerInput) {
-		const { attributesManager } = handlerInput;
-		const attributes = await attributesManager.getPersistentAttributes() || {};
-
-		if (handlerInput.requestEnvelope.request.hasOwnProperty('body')) {
-			if (attributes.hasOwnProperty('optForNotifications')) {
-				attributes.optForNotifications = !attributes.optForNotifications;
-			} else {
-				attributes.optForNotifications = true;
-			}
-		} else {
-			attributes.optForNotifications = false;
-		}
-
-		handlerInput.attributesManager.setPersistentAttributes(attributes);
-		await handlerInput.attributesManager.savePersistentAttributes();
-	},
-};
-
-
 const AudioPlayerEventHandler = {
 	canHandle(handlerInput) {
 		return handlerInput.requestEnvelope.request.type.startsWith('AudioPlayer.');
@@ -98,7 +74,6 @@ const ErrorHandler = {
 };
 
 module.exports = {
-	ProactiveEventHandler,
 	AudioPlayerEventHandler,
 	SessionEndedRequestHandler,
 	ErrorHandler,
